Fix bug in migrations

+++ backend/app/app/models/phrase_this.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DECIMAL
+from sqlalchemy import Column, Integer, String, DECIMAL, BigInteger
 
 from app.db.base_class import Base
 
@@ -10,4 +10,4 @@ class PhraseThis(Base):
     search_volume = Column(Integer, nullable=False, default=0)
     cpc = Column(DECIMAL, nullable=False, default=0.0)
     competition = Column(Integer, nullable=False, default=0)
-    number_results = Column(Integer, nullable=False, default=0)
+    number_results = Column(BigInteger, nullable=False, default=0)
